split test_controllers.py up to test basic and externals separate
